refactor: move review envelope contracts to output crate
