+++ crates/cli/src/output_envelope.rs
@@ -16,12 +16,16 @@ pub use fallow_output::{
     CodeClimateSeverity, CoverageAnalyzeOutput, CoverageAnalyzeSchemaVersion,
     CoverageSetupFileToEdit, CoverageSetupFramework, CoverageSetupMember, CoverageSetupOutput,
     CoverageSetupPackageManager, CoverageSetupRuntimeTarget, CoverageSetupSchemaVersion,
-    CoverageSetupSnippet, DupesOutput, DupesOutputInput, ExplainOutput, GroupByMode, HealthOutput,
-    HealthOutputInput, InspectEvidence, InspectEvidenceScope, InspectEvidenceSection,
-    InspectFileIdentity, InspectIdentity, InspectOutput, InspectSectionStatus,
-    InspectSymbolIdentity, InspectTargetDescriptor, WorkspaceDiagnosticOutput,
-    apply_config_fixable_to_duplicate_exports, build_check_output, build_check_summary,
-    build_dupes_output, build_health_output,
+    CoverageSetupSnippet, DupesOutput, DupesOutputInput, ExplainOutput, GitHubReviewComment,
+    GitHubReviewSide, GitLabReviewComment, GitLabReviewPosition, GitLabReviewPositionType,
+    GroupByMode, HealthOutput, HealthOutputInput, InspectEvidence, InspectEvidenceScope,
+    InspectEvidenceSection, InspectFileIdentity, InspectIdentity, InspectOutput,
+    InspectSectionStatus, InspectSymbolIdentity, InspectTargetDescriptor, MARKER_REGEX_FLAGS_V2,
+    MARKER_REGEX_V2, ReviewCheckConclusion, ReviewComment, ReviewEnvelopeEvent, ReviewEnvelopeMeta,
+    ReviewEnvelopeOutput, ReviewEnvelopeSchema, ReviewEnvelopeSummary, ReviewProvider,
+    WorkspaceDiagnosticOutput, apply_config_fixable_to_duplicate_exports, build_check_output,
+    build_check_summary, build_dupes_output, build_health_output, default_marker_regex,
+    default_marker_regex_flags, is_false,
 };
 use fallow_types::envelope::{ElapsedMs, Meta, SchemaVersion, TelemetryMeta, ToolVersion};
 use fallow_types::output::NextStep;
@@ -222,226 +226,6 @@ pub struct CombinedMeta {
     pub health: Option<Meta>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub telemetry: Option<TelemetryMeta>,
-}
-
-/// Envelope emitted by `fallow --format review-github` / `review-gitlab`.
-#[derive(Debug, Clone, Serialize)]
-#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
-#[cfg_attr(
-    feature = "schema",
-    schemars(title = "fallow --format review-github / review-gitlab")
-)]
-pub struct ReviewEnvelopeOutput {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub event: Option<ReviewEnvelopeEvent>,
-    pub body: String,
-    #[serde(default = "ReviewEnvelopeSummary::empty_default")]
-    pub summary: ReviewEnvelopeSummary,
-    pub comments: Vec<ReviewComment>,
-    #[serde(default = "default_marker_regex")]
-    pub marker_regex: String,
-    #[serde(default = "default_marker_regex_flags")]
-    pub marker_regex_flags: String,
-    pub meta: ReviewEnvelopeMeta,
-}
-
-/// Default for [`ReviewEnvelopeOutput::marker_regex`].
-#[must_use]
-pub fn default_marker_regex() -> String {
-    MARKER_REGEX_V2.to_owned()
-}
-
-/// Default for [`ReviewEnvelopeOutput::marker_regex_flags`].
-#[must_use]
-pub fn default_marker_regex_flags() -> String {
-    MARKER_REGEX_FLAGS_V2.to_owned()
-}
-
-/// Canonical v2 marker-regex literal.
-pub const MARKER_REGEX_V2: &str =
-    r"^<!-- fallow-fingerprint:v2: ((?:[a-z]+:)?[0-9a-f]{16}) -->\s*$";
-
-/// Canonical v2 marker-regex flags.
-pub const MARKER_REGEX_FLAGS_V2: &str = "m";
-
-/// Summary block on [`ReviewEnvelopeOutput`].
-#[derive(Debug, Clone, Serialize, Default)]
-#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
-pub struct ReviewEnvelopeSummary {
-    pub body: String,
-    pub fingerprint: String,
-}
-
-impl ReviewEnvelopeSummary {
-    /// Empty-default factory for [`ReviewEnvelopeOutput::summary`].
-    #[must_use]
-    #[allow(
-        dead_code,
-        reason = "referenced via serde default = \"...\" attr; no direct callsite until Deserialize is derived"
-    )]
-    pub fn empty_default() -> Self {
-        Self::default()
-    }
-}
-
-/// Singleton GitHub review-event marker.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
-#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
-pub enum ReviewEnvelopeEvent {
-    #[serde(rename = "COMMENT")]
-    Comment,
-}
-
-/// Per-line review comment. Schema is an `anyOf` between GitHub and GitLab
-/// shapes; at runtime every entry in a single envelope comes from the same
-/// provider because the envelope is built from one provider's branch in
-/// `crates/cli/src/report/ci/review.rs::render_review_envelope`.
-#[derive(Debug, Clone, Serialize)]
-#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
-#[serde(untagged)]
-pub enum ReviewComment {
-    GitHub(GitHubReviewComment),
-    GitLab(GitLabReviewComment),
-}
-
-/// GitHub pull-request review comment.
-#[derive(Debug, Clone, Serialize)]
-#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
-pub struct GitHubReviewComment {
-    pub path: String,
-    pub line: u32,
-    pub side: GitHubReviewSide,
-    pub body: String,
-    pub fingerprint: String,
-    #[serde(default, skip_serializing_if = "is_false")]
-    pub truncated: bool,
-}
-
-/// Singleton side discriminator for [`GitHubReviewComment::side`].
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
-#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
-pub enum GitHubReviewSide {
-    #[serde(rename = "RIGHT")]
-    Right,
-}
-
-/// GitLab merge-request discussion comment.
-#[derive(Debug, Clone, Serialize)]
-#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
-pub struct GitLabReviewComment {
-    pub body: String,
-    pub position: GitLabReviewPosition,
-    pub fingerprint: String,
-    #[serde(default, skip_serializing_if = "is_false")]
-    pub truncated: bool,
-}
-
-/// Helper for `skip_serializing_if = "is_false"` on `truncated` fields above.
-/// Serde calls `skip_serializing_if` with `&T`, so the reference signature
-/// is dictated by the trait and cannot be changed to pass-by-value. Uses
-/// `#[allow]` rather than `#[expect]` per `.claude/rules/code-quality.md`:
-/// `trivially_copy_pass_by_ref` is a pedantic lint that fires inconsistently
-/// across build configurations (lib vs bin), which would trigger
-/// `unfulfilled_lint_expectations` under `#[expect]`.
-#[must_use]
-#[allow(
-    clippy::trivially_copy_pass_by_ref,
-    reason = "serde's skip_serializing_if requires fn(&T) -> bool"
-)]
-pub fn is_false(value: &bool) -> bool {
-    !*value
-}
-
-/// `position` block inside [`GitLabReviewComment`]. Mirrors the GitLab
-/// merge-request discussion-position API.
-#[derive(Debug, Clone, Serialize)]
-#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
-pub struct GitLabReviewPosition {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub base_sha: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub start_sha: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub head_sha: Option<String>,
-    pub position_type: GitLabReviewPositionType,
-    pub old_path: String,
-    pub new_path: String,
-    pub new_line: u32,
-}
-
-/// Singleton position-type discriminator for [`GitLabReviewPosition`].
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
-#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
-#[serde(rename_all = "lowercase")]
-pub enum GitLabReviewPositionType {
-    Text,
-}
-
-/// `meta` block inside [`ReviewEnvelopeOutput`].
-#[derive(Debug, Clone, Serialize)]
-#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
-pub struct ReviewEnvelopeMeta {
-    pub schema: ReviewEnvelopeSchema,
-    pub provider: ReviewProvider,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub check_conclusion: Option<ReviewCheckConclusion>,
-}
-
-/// Schema-version discriminator for the review envelope.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
-#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
-pub enum ReviewEnvelopeSchema {
-    /// First release of the review envelope format. Historical only; no v1
-    /// emit path remains on the current code. Retained on the enum so a
-    /// future Deserialize derive can still parse v1 captures (e.g. from
-    /// committed snapshots predating the issue #528 migration) without
-    /// erroring on an unknown variant.
-    #[serde(rename = "fallow-review-envelope/v1")]
-    #[allow(
-        dead_code,
-        reason = "kept for forward-compat with v1 historical inputs once Deserialize is derived"
-    )]
-    V1,
-    /// Issue #528 evolution. Adds (1) the [`ReviewEnvelopeOutput::summary`]
-    /// block, (2) [`ReviewEnvelopeOutput::marker_regex`], (3) same-line
-    /// `(path, line)` merging in `comments[]` with a
-    /// `merged:<16-char hash>` primary fingerprint over sorted constituent
-    /// fingerprints (identity shifts whenever the set of constituents
-    /// changes, so the bundled skip-if-fingerprint-exists wrappers
-    /// correctly re-post on content change), (4) UTF-8-safe body
-    /// truncation at the GitLab/GitHub note-size floor (65,536 bytes)
-    /// with paired `truncated: bool` + `<!-- fallow-truncated -->`
-    /// signals, (5) `:v2:`-namespaced marker shape
-    /// (`<!-- fallow-fingerprint:v2: <fingerprint> -->`) preventing v1
-    /// marker collision and user-paste spoofing, and (6) diff-aware
-    /// `position.old_path` for renamed files on GitLab.
-    #[serde(rename = "fallow-review-envelope/v2")]
-    V2,
-}
-
-/// Review-envelope provider tag.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
-#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
-#[serde(rename_all = "lowercase")]
-pub enum ReviewProvider {
-    /// GitHub pull-request review envelope.
-    Github,
-    /// GitLab merge-request discussion envelope.
-    Gitlab,
-}
-
-/// `meta.check_conclusion` for the GitHub review envelope. Maps to the
-/// GitHub Checks API conclusion field.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
-#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
-#[serde(rename_all = "lowercase")]
-pub enum ReviewCheckConclusion {
-    /// No findings.
-    Success,
-    /// Findings but none gated as failure.
-    Neutral,
-    /// At least one finding gated as failure.
-    Failure,
 }
 
 /// Envelope emitted by `fallow ci reconcile-review --format json`. Used by
