+++ crates/output/src/review_envelopes.rs
@@ -1,0 +1,196 @@
+//! Review integration output envelopes.
+
+use serde::Serialize;
+
+/// Envelope emitted by `fallow --format review-github` / `review-gitlab`.
+#[derive(Debug, Clone, Serialize)]
+#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
+#[cfg_attr(
+    feature = "schema",
+    schemars(title = "fallow --format review-github / review-gitlab")
+)]
+pub struct ReviewEnvelopeOutput {
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub event: Option<ReviewEnvelopeEvent>,
+    pub body: String,
+    #[serde(default = "ReviewEnvelopeSummary::empty_default")]
+    pub summary: ReviewEnvelopeSummary,
+    pub comments: Vec<ReviewComment>,
+    #[serde(default = "default_marker_regex")]
+    pub marker_regex: String,
+    #[serde(default = "default_marker_regex_flags")]
+    pub marker_regex_flags: String,
+    pub meta: ReviewEnvelopeMeta,
+}
+
+/// Default for [`ReviewEnvelopeOutput::marker_regex`].
+#[must_use]
+pub fn default_marker_regex() -> String {
+    MARKER_REGEX_V2.to_owned()
+}
+
+/// Default for [`ReviewEnvelopeOutput::marker_regex_flags`].
+#[must_use]
+pub fn default_marker_regex_flags() -> String {
+    MARKER_REGEX_FLAGS_V2.to_owned()
+}
+
+/// Canonical v2 marker-regex literal.
+pub const MARKER_REGEX_V2: &str =
+    r"^<!-- fallow-fingerprint:v2: ((?:[a-z]+:)?[0-9a-f]{16}) -->\s*$";
+
+/// Canonical v2 marker-regex flags.
+pub const MARKER_REGEX_FLAGS_V2: &str = "m";
+
+/// Summary block on [`ReviewEnvelopeOutput`].
+#[derive(Debug, Clone, Serialize, Default)]
+#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
+pub struct ReviewEnvelopeSummary {
+    pub body: String,
+    pub fingerprint: String,
+}
+
+impl ReviewEnvelopeSummary {
+    /// Empty-default factory for [`ReviewEnvelopeOutput::summary`].
+    #[must_use]
+    #[allow(
+        dead_code,
+        reason = "referenced via serde default attr; no direct callsite until Deserialize is derived"
+    )]
+    pub fn empty_default() -> Self {
+        Self::default()
+    }
+}
+
+/// Singleton GitHub review-event marker.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
+#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
+pub enum ReviewEnvelopeEvent {
+    #[serde(rename = "COMMENT")]
+    Comment,
+}
+
+/// Per-line review comment.
+#[derive(Debug, Clone, Serialize)]
+#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
+#[serde(untagged)]
+pub enum ReviewComment {
+    GitHub(GitHubReviewComment),
+    GitLab(GitLabReviewComment),
+}
+
+/// GitHub pull-request review comment.
+#[derive(Debug, Clone, Serialize)]
+#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
+pub struct GitHubReviewComment {
+    pub path: String,
+    pub line: u32,
+    pub side: GitHubReviewSide,
+    pub body: String,
+    pub fingerprint: String,
+    #[serde(default, skip_serializing_if = "is_false")]
+    pub truncated: bool,
+}
+
+/// Singleton side discriminator for [`GitHubReviewComment::side`].
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
+#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
+pub enum GitHubReviewSide {
+    #[serde(rename = "RIGHT")]
+    Right,
+}
+
+/// GitLab merge-request discussion comment.
+#[derive(Debug, Clone, Serialize)]
+#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
+pub struct GitLabReviewComment {
+    pub body: String,
+    pub position: GitLabReviewPosition,
+    pub fingerprint: String,
+    #[serde(default, skip_serializing_if = "is_false")]
+    pub truncated: bool,
+}
+
+/// Helper for `skip_serializing_if = "is_false"` on `truncated` fields.
+#[must_use]
+#[allow(
+    clippy::trivially_copy_pass_by_ref,
+    reason = "serde's skip_serializing_if requires fn(&T) -> bool"
+)]
+pub fn is_false(value: &bool) -> bool {
+    !*value
+}
+
+/// `position` block inside [`GitLabReviewComment`].
+#[derive(Debug, Clone, Serialize)]
+#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
+pub struct GitLabReviewPosition {
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub base_sha: Option<String>,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub start_sha: Option<String>,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub head_sha: Option<String>,
+    pub position_type: GitLabReviewPositionType,
+    pub old_path: String,
+    pub new_path: String,
+    pub new_line: u32,
+}
+
+/// Singleton position-type discriminator for [`GitLabReviewPosition`].
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
+#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
+#[serde(rename_all = "lowercase")]
+pub enum GitLabReviewPositionType {
+    Text,
+}
+
+/// `meta` block inside [`ReviewEnvelopeOutput`].
+#[derive(Debug, Clone, Serialize)]
+#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
+pub struct ReviewEnvelopeMeta {
+    pub schema: ReviewEnvelopeSchema,
+    pub provider: ReviewProvider,
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub check_conclusion: Option<ReviewCheckConclusion>,
+}
+
+/// Schema-version discriminator for the review envelope.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
+#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
+pub enum ReviewEnvelopeSchema {
+    /// Historical first release of the review envelope format.
+    #[serde(rename = "fallow-review-envelope/v1")]
+    #[allow(
+        dead_code,
+        reason = "kept for forward-compat with v1 historical inputs once Deserialize is derived"
+    )]
+    V1,
+    /// Issue #528 review envelope format.
+    #[serde(rename = "fallow-review-envelope/v2")]
+    V2,
+}
+
+/// Review-envelope provider tag.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
+#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
+#[serde(rename_all = "lowercase")]
+pub enum ReviewProvider {
+    /// GitHub pull-request review envelope.
+    Github,
+    /// GitLab merge-request discussion envelope.
+    Gitlab,
+}
+
+/// `meta.check_conclusion` for the GitHub review envelope.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
+#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
+#[serde(rename_all = "lowercase")]
+pub enum ReviewCheckConclusion {
+    /// No findings.
+    Success,
+    /// Findings but none gated as failure.
+    Neutral,
+    /// At least one finding gated as failure.
+    Failure,
+}
